+++ example/document_example.dart
@@ -5,7 +5,7 @@ Future<void> main(List<String> args) async {
   final dm = DocumentModel(c);
 
   try {
-    DbResponse r = await dm.getDoc('db', 'some_id');
+    DbResponse r = await dm.deleteDoc('ddds', 'asdf', '1-47abcaa8abbf8ce6ce772d92e423e24d');
     print('$r - success');
   } on CouchDbException catch (e) {
     print('$e - error');