+++ lib/src/clients/couchdb_web_client.dart
@@ -68,5 +68,4 @@
 //   String username;
 //   String password;
 
-
 // }