+++ lib/src/models/base/base_model.dart
@@ -1,8 +1,7 @@
 import '../../clients/base/couchdb_base_client.dart';
 
-/// Base model that hold connected [client] 
+/// Base model that hold connected [client]
 abstract class BaseModel {
-
   /// Creates instanse of [BaseModel] with given [client]
   BaseModel(this.client);