+++ lib/src/models/base/local_document_base_model.dart
@@ -5,13 +5,10 @@ import 'base_model.dart';
 
 /// Class is under heavy development - don't use it for implementing your own methods!
 abstract class LocalDocumentBaseModel extends BaseModel {
+  LocalDocumentBaseModel(CouchDbBaseClient client) : super(client);
 
-  LocalDocumentBaseModel(CouchDbBaseClient client): super(client);
-
-  Future<String> localDocs(
-    String dbName,
-    {
-      bool conflicts = false,
+  Future<String> localDocs(String dbName,
+      {bool conflicts = false,
       bool descending = false,
       String endKey,
       String endKeyDocId,
@@ -23,13 +20,9 @@ abstract class LocalDocumentBaseModel extends BaseModel {
       int skip = 0,
       String startKey,
       String startKeyDocId,
-      bool updateSeq = false
-    }
-  );
-  Future<String> localDocsWithKeys(
-    String dbName,
-    {
-      @required List<String> keys,
+      bool updateSeq = false});
+  Future<String> localDocsWithKeys(String dbName,
+      {@required List<String> keys,
       bool conflicts = false,
       bool descending = false,
       String endKey,
@@ -41,9 +34,7 @@ abstract class LocalDocumentBaseModel extends BaseModel {
       int skip = 0,
       String startKey,
       String startKeyDocId,
-      bool updateSeq = false
-    }
-  );
+      bool updateSeq = false});
   Future<String> getLocalDoc(String dbName, String docId);
   Future<String> putLocalDoc(String dbName, String docId);
   Future<String> deleteLocalDoc(String dbName, String docId);