+++ lib/src/models/base/server_base_model.dart
@@ -5,42 +5,43 @@ import 'base_model.dart';
 
 /// Class is under heavy development - don't use it for implementing your own methods!
 abstract class ServerBaseModel extends BaseModel {
-
-  ServerBaseModel(CouchDbBaseClient client): super(client);
+  ServerBaseModel(CouchDbBaseClient client) : super(client);
 
   Future<String> serverInfo();
   Future<List<String>> activeTasks();
   Future<List<String>> allDbs();
   Future<List<String>> dbsInfo(List<String> keys);
-  Future<List<String>> clusterSetup({
-    List<String> ensureDbsExist = const <String>['_users','_replicator','_global_changes']
-  });
+  Future<List<String>> clusterSetup(
+      {List<String> ensureDbsExist = const <String>[
+        '_users',
+        '_replicator',
+        '_global_changes'
+      ]});
   // Cluster setup API with POST will be soon
-  Future<List<String>> dbUpdates({
-    String feed = 'normal',
-    int timeout = 60,
-    int heartbeat = 60000,
-    String since
-  });
+  Future<List<String>> dbUpdates(
+      {String feed = 'normal',
+      int timeout = 60,
+      int heartbeat = 60000,
+      String since});
   Future<String> membership();
-  Future<String> replicate({
-    bool cancel,
-    bool continuous,
-    bool createTarget,
-    List<String> docIds,
-    String filterFunJS,
-    String proxy,
-    Object source,
-    Object target
-  });
-  Future<String> schedulerJobs({ int limit, int skip });
-  Future<String> schedulerDocs({ int limit, int skip });
-  Future<String> schedulerDocsWithReplicatiorDbName({ @required String replicator, int limit, int skip });
+  Future<String> replicate(
+      {bool cancel,
+      bool continuous,
+      bool createTarget,
+      List<String> docIds,
+      String filterFunJS,
+      String proxy,
+      Object source,
+      Object target});
+  Future<String> schedulerJobs({int limit, int skip});
+  Future<String> schedulerDocs({int limit, int skip});
+  Future<String> schedulerDocsWithReplicatiorDbName(
+      {@required String replicator, int limit, int skip});
   Future<String> schedulerDocsWithDocId(String replicator, String docId);
-  Future<String> nodeStats({ String nodeName = '_local' });
-  Future<String> systemStatsForNode({ String nodeName = '_local' });
+  Future<String> nodeStats({String nodeName = '_local'});
+  Future<String> systemStatsForNode({String nodeName = '_local'});
   // _utils will be soon
   Future<String> up();
-  Future<List<String>> uuids({ int count = 1 });
+  Future<List<String>> uuids({int count = 1});
   // favicon ?
 }