+++ lib/src/models/document_model.dart
@@ -8,16 +8,12 @@ import 'base/document_base_model.dart';
 
 /// Class that implements methods for create, read, update and delete documents within a database
 class DocumentModel extends DocumentBaseModel {
-
   /// Create DocumentModel by accepting web-based or server-based client
   DocumentModel(CouchDbBaseClient client) : super(client);
 
   @override
-  Future<DbResponse> docInfo(
-    String dbName,
-    String docId,
-    {
-      Map<String, String> headers,
+  Future<DbResponse> docInfo(String dbName, String docId,
+      {Map<String, String> headers,
       bool attachments = false,
       bool attEncodingInfo = false,
       List<String> attsSince,
@@ -29,15 +25,14 @@ class DocumentModel extends DocumentBaseModel {
       Object openRevs,
       String rev,
       bool revs = false,
-      bool revsInfo = false
-    }
-  ) async {
+      bool revsInfo = false}) async {
     DbResponse result;
 
-    final path = '$dbName/$docId?attachments=$attachments&att_encoding_info=$attEncodingInfo&'
-      '${includeNonNullParam('atts_since', attsSince)}&conflicts=$conflicts&deleted_conflicts=$deletedConflicts&'
-      'latest=$latest&local_seq=$localSeq&meta=$meta&${includeNonNullParam('open_revs', openRevs)}&'
-      '${includeNonNullParam('rev', rev)}&revs=$revs&revs_info=$revsInfo';
+    final path =
+        '$dbName/$docId?attachments=$attachments&att_encoding_info=$attEncodingInfo&'
+        '${includeNonNullParam('atts_since', attsSince)}&conflicts=$conflicts&deleted_conflicts=$deletedConflicts&'
+        'latest=$latest&local_seq=$localSeq&meta=$meta&${includeNonNullParam('open_revs', openRevs)}&'
+        '${includeNonNullParam('rev', rev)}&revs=$revs&revs_info=$revsInfo';
 
     try {
       result = await client.head(path, reqHeaders: headers);
@@ -48,11 +43,8 @@ class DocumentModel extends DocumentBaseModel {
   }
 
   @override
-  Future<DbResponse> getDoc(
-    String dbName,
-    String docId,
-    {
-      Map<String, String> headers,
+  Future<DbResponse> getDoc(String dbName, String docId,
+      {Map<String, String> headers,
       bool attachments = false,
       bool attEncodingInfo = false,
       List<String> attsSince,
@@ -64,15 +56,14 @@ class DocumentModel extends DocumentBaseModel {
       Object openRevs,
       String rev,
       bool revs = false,
-      bool revsInfo = false
-    }
-  ) async {
+      bool revsInfo = false}) async {
     DbResponse result;
 
-    final path = '$dbName/$docId?attachments=$attachments&att_encoding_info=$attEncodingInfo&'
-      '${includeNonNullParam('atts_since', attsSince)}&conflicts=$conflicts&deleted_conflicts=$deletedConflicts&'
-      'latest=$latest&local_seq=$localSeq&meta=$meta&${includeNonNullParam('open_revs', openRevs)}&'
-      '${includeNonNullParam('rev', rev)}&revs=$revs&revs_info=$revsInfo';
+    final path =
+        '$dbName/$docId?attachments=$attachments&att_encoding_info=$attEncodingInfo&'
+        '${includeNonNullParam('atts_since', attsSince)}&conflicts=$conflicts&deleted_conflicts=$deletedConflicts&'
+        'latest=$latest&local_seq=$localSeq&meta=$meta&${includeNonNullParam('open_revs', openRevs)}&'
+        '${includeNonNullParam('rev', rev)}&revs=$revs&revs_info=$revsInfo';
 
     try {
       result = await client.get(path, reqHeaders: headers);
@@ -83,88 +74,71 @@ class DocumentModel extends DocumentBaseModel {
   }
 
   @override
-  Future<String> insertDoc(
-    String dbName,
-    String docId,
-    {
+  Future<DbResponse> insertDoc(
+      String dbName, String docId, Map<String, Object> body,
+      {Map<String, String> headers,
       String rev,
       String batch,
-      bool newEdits = true
-    }
-  ) async {
+      bool newEdits = true}) async {
+    DbResponse result;
 
+    final path =
+        '$dbName/$docId?new_edits=$newEdits&${includeNonNullParam('rev', rev)}&'
+        '${includeNonNullParam('batch', batch)}';
+
+    try {
+      result = await client.put(path, reqHeaders: headers, body: body);
+    } on CouchDbException {
+      rethrow;
+    }
+    return result;
   }
 
   @override
-  Future<String> deleteDoc(
-    String dbName,
-    String docId,
-    String rev,
-    {
-      String batch
-    }
-  ) async {
+  Future<DbResponse> deleteDoc(String dbName, String docId, String rev,
+      {Map<String, String> headers, String batch}) async {
+    DbResponse result;
 
+    final path =
+        '$dbName/$docId?rev=$rev&${includeNonNullParam('batch', batch)}';
+
+    try {
+      result = await client.delete(path, reqHeaders: headers);
+    } on CouchDbException {
+      rethrow;
+    }
+    return result;
   }
 
   @override
-  Future<String> copyDoc(
-    String dbName,
-    String docId,
-    {
-      String rev,
-      String batch
-    }
-  ) async {
+  Future<DbResponse> copyDoc(String dbName, String docId,
+      {Map<String, String> headers, String rev, String batch}) async {
+    DbResponse result;
 
+    final path = '$dbName/$docId?${includeNonNullParam('rev', rev)}&'
+        '${includeNonNullParam('batch', batch)}';
+
+    try {
+      result = await client.copy(path, reqHeaders: headers);
+    } on CouchDbException {
+      rethrow;
+    }
+    return result;
   }
 
   @override
-  Future<String> attachmentInfo(
-    String dbName,
-    String docId,
-    String attName,
-    {
-      String rev
-    }
-  ) async {
-
-  }
+  Future<String> attachmentInfo(String dbName, String docId, String attName,
+      {String rev}) async {}
 
   @override
-  Future<String> getAttachment(
-    String dbName,
-    String docId,
-    String attName,
-    {
-      String rev
-    }
-  ) async {
-
-  }
+  Future<String> getAttachment(String dbName, String docId, String attName,
+      {String rev}) async {}
 
   @override
-  Future<String> insertAttachment(
-    String dbName,
-    String docId,
-    String attName,
-    {
-      String rev
-    }
-  ) async {
-
-  }
+  Future<String> insertAttachment(String dbName, String docId, String attName,
+      {String rev}) async {}
 
   @override
-  Future<String> deleteAttachment(
-    String dbName,
-    String docId,
-    String attName,
-    {
-      @required String rev,
-      String batch
-    }
-  ) async {
-
-  }
+  Future<String> deleteAttachment(String dbName, String docId, String attName,
+      {@required String rev, String batch}) async {}
 }