+++ lib/src/utils/includer_path.dart
@@ -1,2 +1,3 @@
 /// Method for including only non-null parameter to path
-String includeNonNullParam(String name, Object value) => value != null ? '$name=$value' : '';
+String includeNonNullParam(String name, Object value) =>
+    value != null ? '$name=$value' : '';