+++ test/couchdb_test.dart
@@ -2,12 +2,8 @@ import 'package:test/test.dart';
 
 void main() {
   group('A group of tests', () {
+    setUp(() {});
 
-    setUp(() {
-      
-    });
-
-    test('First Test', () {
-    });
+    test('First Test', () {});
   });
 }