+++ CHANGELOG.md
@@ -1,5 +1,16 @@
-# 0.0.2
+# 0.0.3
+
+- Implements `insertDoc()`, `deleteDoc()` and `copyDoc()` methods of `DocumentModel`.
+- All methods of `CouchDb(Server/Web)Client` can set custom headers.
+
+## 0.0.2
 
 - Implementats `docInfo()` and `getDoc()` methods of `DocumentModel`.
 - Add documentation to `DatabaseModel` methods.
 - Add opportunity to add headers to `head()` method of `CouchDb(Server/Web)Client`
+
+## 0.0.1
+
+- Initial version.
+- Implement `CouchDbServerClient` for interacting with CouchDB.
+- Implements `DatabaseBaseModel` methods.